+++ freemarker-generator-tools/src/main/java/org/apache/freemarker/generator/tools/jsonpath/JsonPathTool.java
@@ -24,13 +24,24 @@ import org.apache.freemarker.generator.base.datasource.DataSource;
 
 import java.io.IOException;
 import java.io.InputStream;
+import java.util.Collection;
+import java.util.List;
+import java.util.stream.Collectors;
 
 public class JsonPathTool {
 
-    public DocumentContext parse(DataSource dataSource) throws IOException {
+    public DocumentContext parse(DataSource dataSource) {
         try (InputStream is = dataSource.getUnsafeInputStream()) {
             return JsonPath.using(configuration()).parse(is);
+        } catch (IOException e) {
+            throw new RuntimeException("Failed to parse data source:" + dataSource, e);
         }
+    }
+
+    public List<DocumentContext> parse(Collection<DataSource> dataSources) {
+        return dataSources.stream()
+                .map(this::parse)
+                .collect(Collectors.toList());
     }
 
     public DocumentContext parse(String json) {